Exception handling added

+++ src/main/java/com/example/demo/Service/ProductService.java
@@ -1,5 +1,7 @@
 package com.example.demo.Service;
 
+import com.example.demo.Exception.InvalidDataException;
+import com.example.demo.Exception.ProductNotFoundException;
 import com.example.demo.Model.Product;
 import com.example.demo.Repository.ProductRepository;
 import org.springframework.stereotype.Service;
@@ -20,14 +22,35 @@ public class ProductService {
     }
 
     public Product getProductById(Long id) {
-        return productRepository.findById(id).orElseThrow(() -> new RuntimeException("Product not found"));
+        if (productRepository.findById(id).isEmpty()) {
+            throw new ProductNotFoundException("There is no product with the id that you entered");
+        }
+
+            return productRepository.findById(id).get();
+
     }
 
     public Product createProduct(Product product) {
+        if(product.getPrice() <= 0){
+            throw new InvalidDataException("The price should be greater than zero.");
+        }
+
+        if(product.getName().isEmpty() || product.getDescription().isEmpty() ){
+            throw new InvalidDataException("The name and description cannot be empty.");
+        }
         return productRepository.save(product);
     }
 
     public Product updateProduct(Long id, Product updatedProduct) {
+
+        if(updatedProduct.getPrice() <= 0){
+            throw new InvalidDataException("The price should be greater than zero.");
+        }
+
+        if(updatedProduct.getName().isEmpty() || updatedProduct.getDescription().isEmpty() ){
+            throw new InvalidDataException("The name and description cannot be empty.");
+        }
+
         Product product = getProductById(id);
         product.setName(updatedProduct.getName());
         product.setDescription(updatedProduct.getDescription());
@@ -36,6 +59,16 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+
+
+        if(productRepository.findById(id).isEmpty()){
+            throw new ProductNotFoundException(
+                    "There is no product with the id that you entered");
+        }
+
+            productRepository.deleteById(id);
+
+        }
+
     }
-}
+
